AI_Workflow360/ai-server/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
transformers==4.41.0
torch==2.3.0
scikit-learn==1.5.0
sentence-transformers==3.0.0
xgboost==2.0.3
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
